requirements.txt
pandas
numpy
websocket-client
simplejson